Feat [#5] 서비스를 조금 더 구체화 아직 skltn 단계
