--- morib/src/main/java/org/morib/server/api/timerView/service/fetch/FetchTimerService.java
@@ -1,5 +1,0 @@
-package org.morib.server.api.timerView.service.fetch;
-
-public interface FetchTimerService {
-    void fetch();
-}
