--- morib/src/main/java/org/morib/server/api/timerView/service/fetch/FetchTimerServiceImpl.java
@@ -1,14 +1,0 @@
-package org.morib.server.api.timerView.service.fetch;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class FetchTimerServiceImpl implements FetchTimerService{
-
-    @Override
-    public void fetch() {
-
-    }
-}
