+++ morib/src/main/java/org/morib/server/api/timerView/service/fetch/todo/FetchTodoService.java
@@ -1,0 +1,7 @@
+package org.morib.server.api.timerView.service.fetch.todo;
+
+import java.time.LocalDate;
+
+public interface FetchTodoService {
+    void fetch(LocalDate targetDate);
+}
