+++ morib/src/main/java/org/morib/server/api/timerView/service/fetch/todo/FetchTodoServiceImpl.java
@@ -1,0 +1,23 @@
+package org.morib.server.api.timerView.service.fetch.todo;
+
+import java.time.LocalDate;
+import lombok.RequiredArgsConstructor;
+import org.morib.server.domain.todo.application.TodoGateway;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class FetchTodoServiceImpl implements FetchTodoService {
+
+    private final TodoGateway todoGateway;
+
+    @Override
+    public void fetch() {
+
+    }
+
+    @Override
+    public void todoFetch(LocalDate targetDate) {
+
+    }
+}
