--- morib/src/main/java/org/morib/server/api/timerView/service/stop/StopTimerService.java
@@ -1,5 +1,0 @@
-package org.morib.server.api.timerView.service.stop;
-
-public interface StopTimerService {
-    void stop();
-}
