--- morib/src/main/java/org/morib/server/api/timerView/service/stop/StopTimerServiceImpl.java
@@ -1,14 +1,0 @@
-package org.morib.server.api.timerView.service.stop;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class StopTimerServiceImpl implements StopTimerService{
-
-    @Override
-    public void stop() {
-
-    }
-}
