+++ morib/src/main/java/org/morib/server/api/timerView/service/stop/timer/StopTimerService.java
@@ -1,0 +1,5 @@
+package org.morib.server.api.timerView.service.stop.timer;
+
+public interface StopTimerService {
+    void stop();
+}
